+++ packages/@cldcvr/flow-lineage/CHANGELOG.md
@@ -11,6 +11,9 @@
 
 - `rightClick` event support added in node schema.
 
+### Improvements
+- `isChildrenVisible` changed to `hideChildren` with default value `true.`
+
 ## [1.0.2] - 2023-01-4
 
 ### Improvements